implement several drivers

+++ masteringMCU/stm32f4xx_drivers/drivers/Inc/stm32f446xx_gpio_driver.h
@@ -143,7 +143,7 @@ void GPIO_DeInit(GPIO_RegDef_t *pGPIOx);
  * data read and write
  */
 
-uint32_t GPIO_ReadFromInputPin(GPIO_RegDef_t *pGPIOx, uint32_t PinNumber);
+uint16_t GPIO_ReadFromInputPin(GPIO_RegDef_t *pGPIOx, uint16_t PinNumber);
 uint16_t GPIO_ReadFromInputPort(GPIO_RegDef_t *pGPIOx);
 void GPIO_WriteToOutputPin(GPIO_RegDef_t *pGPIOx, uint32_t PinNumber, uint32_t Value);
 void GPIO_WriteToOutputPort(GPIO_RegDef_t *pGPIOx, uint16_t Value);
